Contents component: make sure that our simulation property editor is given the raw version of our File object.
Otherwise, it will be a Proxy version of it and this means that, in our Web app, `this._sedDocument.simulation(index).constructor.name` will not work. To make it work, we would need to do `vue.toRaw(this)._sedDocument.simulation(index).constructor.name`... and the same wherever we use `this`.