CD: use an architecture-based channel to publish our `.yml` files.
A `.yml` file is generated by `electron-builder`. It is used for auto updating and, on Linux, it all works fine for both Intel and ARM since `electron-builder` generates two different `.yml` files (`latest-linux.yml` and `latest-linux-arm64.yml`), but this is not the case on Windows and macOS. There, there is only one `.yml` file: `latest.yml` for Windows and `latest-mac.yml`.

One way to "fix" the issue is therefore to use an architecture-based channel resulting in the following `.yml` files:
- Windows Intel: `latest-x64.yml`
- Windows ARM: `latest-arm64.yml`
- Linux Intel: `latest-x64-linux.yml`
- Linux ARM: `latest-arm64-linux-arm64.yml`
- macOS Intel: `latest-x64-mac.yml`
- macOS ARM: `latest-arm64-mac.yml`

Filenames are anything but consistent, but then again they were not consistent to start with!

. On Windows and macOS,